src/extractors/table_extractor.py
from .base_extractor import BaseExtractor
import fitz  # PyMuPDF
import pandas as pd
import uuid
import os

class TableExtractor(BaseExtractor):
    def __init__(self, config):
        self.config = config
        self.min_rows = config.get('min_rows', 2)
        self.min_cols = config.get('min_cols', 2)
        
    def extract(self, pdf_path):
        try:
            # In a real implementation, you would use a library like tabula-py
            # For simplicity, we'll just return a placeholder
            doc = fitz.open(pdf_path)
            table_data = []
            
            for page_num, page in enumerate(doc):
                # This is a simplified approach - in a real implementation
                # you would use more sophisticated table detection
                text = page.get_text("dict")
                
                # Simple heuristic: if we have blocks with similar y-coordinates,
                # they might be part of a table
                if len(text["blocks"]) > self.min_rows:
                    table_data.append({
                        'id': str(uuid.uuid4()),
                        'page': page_num,
                        'text': f"Table found on page {page_num+1}",
                        'type': 'table'
                    })
                
            return table_data
        except Exception as e:
            print(f"Error extracting tables: {str(e)}")
            return []