requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
numpy>=1.24.0
PyMuPDF>=1.22.0
pandas>=1.5.3
Pillow>=9.5.0
pyyaml>=6.0
sentence-transformers>=2.2.2